Simplify target handling and ignores
Check for ignored packages at the very end. Less efficient, but reads
better and provides a consistent experience for installs, upgrades and
mixed commands.

Currently there's no test coverage on this feature, but at the unit
level feels too invasive, maybe an end-to-end integration test could
confirm that ignored packages are correctly ignored (as well as
prevention of duplicate targets).