Refactor download, simplify specs
To support the refactor of download we needed to simplify the handling
of the Config object during specs. We now always return an immutable
copy of the DEFAULTS for all calls. If a specific spec (like Options)
needs to modify the config, it must set the (new) mutable attribute to
true on the Config mock.